+++ src/cli/commands/mc-version-help.ts
@@ -13,15 +13,15 @@ export default (program: CommanderStatic) => {
 
     program.on("--help", () => {
         log(`\n  Documentation:\n`);
-        log(`    the ${magenta("magenta colored commands *")} require service credentials.`);
+        log(`    the ${magenta("magenta colored commands *")} require app or service credentials.`);
         log(`    the ${cyan("cyan colored commands ")}require mindconnectlib (agent) credentials`);
         log(`    the ${blue("blue colored commands @")} use analytical functions of MindSphere`);
         log(`    the ${green("green colored commands #")} are used as setup and utility commands`);
-        log(`    the service credentials should only be used in secure environments for setup tasks`);
+        log(`    the app or service credentials should only be used in secure environments`);
         log(`    Full documentation: ${cyan("https://opensource.mindsphere.io")}\n`);
     });
 
-    program.on("command:*", function() {
+    program.on("command:*", function () {
         console.error(red(`Invalid command: ${program.args.join(" ")}`));
         console.error("Use mc --help for a list of available commands");
         process.exit(1);